port the remaining AdcMatrix tests
